+++ .github/workflows/code-quality.yml
@@ -19,28 +19,19 @@ jobs:
       uses: actions/setup-node@v4
       with:
         node-version: '20.x'
-        cache: 'yarn'
+        cache: 'npm'
         
     - name: Install dependencies
-      run: yarn install --frozen-lockfile
+      run: npm ci
       
     - name: Type checking
-      run: yarn type-check
+      run: npm run type-check
       
     - name: Linting (ESLint)
-      run: yarn lint
+      run: npm run lint
       
     - name: Code formatting check (Prettier)
-      run: yarn format --check
-      continue-on-error: true
+      run: npm run format -- --check
       
-    - name: Check for security vulnerabilities
-      run: yarn audit --level moderate
-      continue-on-error: true
-      
-    - name: Quality Summary
-      run: |
-        echo "## Code Quality Summary" >> $GITHUB_STEP_SUMMARY
-        echo "✅ Type checking completed" >> $GITHUB_STEP_SUMMARY
-        echo "✅ Linting completed" >> $GITHUB_STEP_SUMMARY
-        echo "✅ Security audit completed" >> $GITHUB_STEP_SUMMARY
+    - name: Build check
+      run: npm run build
